Remove requirements file
Ideally there should be a single source of truth -- since some things are available only on conda and not on pypi, we have 1.5 sources (pyproject.toml and environment.yml), but these still supersede requirements.txt.

I went through and removed every requirement that wasn't either (a) didn't appear directly anywhere in the project, or (b) was already in our other requirements lists. All that remained was `atomistics`. This module is used directly in `pyiron_nodes.atomistic.property.elastic`, so I added it to the dependencies. (We got away without explicitly requiring it since it came with `pyiron_atomistics` anyhow, but any package that is directly referenced should appear explicitly in the dependencies.)